# cmd/vectodblite_cluster/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  main.Health:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
  main.ReqAcquire:
    properties:
      dbID:
        type: integer
      nodeAddr:
        type: string
    type: object
  main.ReqAdd:
    properties:
      dbID:
        type: integer
      xb:
        items:
          type: number
        type: array
      xid:
        type: integer
    type: object
  main.ReqRelease:
    properties:
      dbID:
        type: integer
    type: object
  main.ReqSearch:
    properties:
      dbID:
        type: integer
      xq:
        items:
          type: number
        type: array
    type: object
  main.RspAcquire:
    properties:
      dbID:
        type: integer
      err:
        type: string
      nodeAddr:
        type: string
    type: object
  main.RspAdd:
    properties:
      err:
        type: string
      xid:
        type: integer
    type: object
  main.RspRelease:
    properties:
      dbID:
        type: integer
      err:
        type: string
    type: object
  main.RspSearch:
    properties:
      distance:
        type: number
      err:
        type: string
      xid:
        type: integer
    type: object
  main.Status:
    properties:
      status:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact:
    email: yuzhichang@gmail.com
    name: API Support
    url: https://github.com/infinivision/vectodb/issues
  description: This is a VectoDBLite cluster server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VectoDBLite Cluster API
  version: "1.0"
paths:
  /api/v1/add:
    post:
      consumes:
      - application/json
      description: Add a vector to the given vectodblite
      parameters:
      - description: ReqAdd. If xid is 0 or ^uint64(0), the cluster will generate
          one.
        in: body
        name: add
        required: true
        schema:
          $ref: '#/definitions/main.ReqAdd'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: RspAdd
          schema:
            $ref: '#/definitions/main.RspAdd'
            type: object
        "301":
          description: redirection
        "400": {}
  /api/v1/search:
    post:
      consumes:
      - application/json
      description: Search a vector in the given vectodblite
      parameters:
      - description: ReqSearch
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/main.ReqSearch'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: RspSearch
          schema:
            $ref: '#/definitions/main.RspSearch'
            type: object
        "301":
          description: redirection
        "400": {}
  /health:
    get:
      description: Eureka healthCheckUrl.
      produces:
      - application/json
      responses:
        "200":
          description: Health
          schema:
            $ref: '#/definitions/main.Health'
            type: object
  /mgmt/v1/acquire:
    post:
      consumes:
      - application/json
      description: Assocaite a vectodblite with the given node. Only the leader node
        supports this API.
      parameters:
      - description: ReqAcquire
        in: body
        name: add
        required: true
        schema:
          $ref: '#/definitions/main.ReqAcquire'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: RspAcquire
          schema:
            $ref: '#/definitions/main.RspAcquire'
            type: object
        "301":
          description: redirection
        "400": {}
  /mgmt/v1/release:
    post:
      consumes:
      - application/json
      description: De-associate a vectodblite with this node.
      parameters:
      - description: ReqRelease
        in: body
        name: add
        required: true
        schema:
          $ref: '#/definitions/main.ReqRelease'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: RspRelease
          schema:
            $ref: '#/definitions/main.RspRelease'
            type: object
        "301":
          description: redirection
        "400": {}
  /status:
    get:
      description: Eureka statusPageUrl.
      produces:
      - application/json
      responses:
        "200":
          description: Status
          schema:
            $ref: '#/definitions/main.Status'
            type: object
swagger: "2.0"